Implement sign out from current space
