+++ lib/ui/shared/appbar_drawer/drawer/bloc/app_drawer_event.dart
@@ -10,5 +10,4 @@ class ChangeSpaceEvent extends DrawerEvents {
   ChangeSpaceEvent(this.space);
 }
 
-class SignOutEvent extends DrawerEvents {
-}
+class SignOutWithCurrentSpaceEvent extends DrawerEvents {}
