+++ lib/ui/space/join_space/bloc/join_space_bloc.dart
@@ -5,6 +5,7 @@ import 'package:projectunity/data/core/exception/error_const.dart';
 import 'package:projectunity/data/model/employee/employee.dart';
 import 'package:projectunity/data/provider/user_state.dart';
 import 'package:projectunity/data/services/account_service.dart';
+import 'package:projectunity/data/services/auth_service.dart';
 import 'package:projectunity/ui/space/join_space/bloc/join_space_event.dart';
 import 'package:projectunity/ui/space/join_space/bloc/join_space_state.dart';
 import '../../../../data/core/utils/bloc_status.dart';
@@ -21,14 +22,16 @@ class JoinSpaceBloc extends Bloc<JoinSpaceEvents, JoinSpaceState> {
   final SpaceService _spaceService;
   final InvitationService _invitationService;
   final AccountService accountService;
+  final AuthService _authService;
   late final List<Invitation> invitations;
 
   JoinSpaceBloc(this._invitationService, this._spaceService, this._userManager,
-      this.accountService, this._employeeService)
+      this.accountService, this._employeeService, this._authService)
       : super(const JoinSpaceState()) {
     on<JoinSpaceInitialFetchEvent>(_init);
     on<SelectSpaceEvent>(_selectSpace);
     on<JoinRequestedSpaceEvent>(_joinRequestedSpace);
+    on<SignOutEvent>(_signOut);
   }
 
   String get userEmail => _userManager.userEmail ?? "unknown";
@@ -116,5 +119,21 @@ class JoinSpaceBloc extends Bloc<JoinSpaceEvents, JoinSpaceState> {
   Invitation deleteInvitation(String spaceId) {
     return invitations
         .firstWhere((invitation) => invitation.spaceId == spaceId);
+  }
+
+  Future<void> _signOut(
+      SignOutEvent event, Emitter<JoinSpaceState> emit) async {
+    emit(state.copyWith(signOutStatus: Status.loading));
+    try {
+      bool isLogOut = await _authService.signOutWithGoogle();
+      if (isLogOut) {
+        await _userManager.removeAll();
+        emit(state.copyWith(signOutStatus: Status.success));
+      } else {
+        throw Exception(signOutError);
+      }
+    } on Exception {
+      emit(state.copyWith(error: signOutError, signOutStatus: Status.error));
+    }
   }
 }
