+++ lib/ui/space/join_space/bloc/join_space_state.dart
@@ -5,6 +5,7 @@ import '../../../../data/model/space/space.dart';
 class JoinSpaceState extends Equatable {
   final Status fetchSpaceStatus;
   final Status selectSpaceStatus;
+  final Status signOutStatus;
   final List<Space> ownSpaces;
   final List<Space> requestedSpaces;
   final String? error;
@@ -12,25 +13,34 @@ class JoinSpaceState extends Equatable {
   const JoinSpaceState({
     this.selectSpaceStatus = Status.initial,
     this.fetchSpaceStatus = Status.initial,
+    this.signOutStatus = Status.initial,
     this.ownSpaces = const [],
     this.requestedSpaces = const [],
     this.error,
   });
 
   JoinSpaceState copyWith(
-          {Status? fetchSpaceStatus,
+          {Status? signOutStatus,
+          Status? fetchSpaceStatus,
           Status? selectSpaceStatus,
           List<Space>? ownSpaces,
           List<Space>? requestedSpaces,
           String? error}) =>
       JoinSpaceState(
           error: error,
+          signOutStatus: signOutStatus ?? this.signOutStatus,
           selectSpaceStatus: selectSpaceStatus ?? this.selectSpaceStatus,
           fetchSpaceStatus: fetchSpaceStatus ?? this.fetchSpaceStatus,
           requestedSpaces: requestedSpaces ?? this.requestedSpaces,
           ownSpaces: ownSpaces ?? this.ownSpaces);
 
   @override
-  List<Object?> get props =>
-      [fetchSpaceStatus, ownSpaces, requestedSpaces, error, selectSpaceStatus];
+  List<Object?> get props => [
+        fetchSpaceStatus,
+        ownSpaces,
+        requestedSpaces,
+        error,
+        selectSpaceStatus,
+        signOutStatus
+      ];
 }
