+++ test/unit_test/shared/drawer/drawer_test.mocks.dart
@@ -3,21 +3,20 @@
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i9;
-import 'dart:ui' as _i11;
+import 'dart:async' as _i8;
+import 'dart:ui' as _i10;
 
 import 'package:cloud_firestore/cloud_firestore.dart' as _i2;
-import 'package:firebase_auth/firebase_auth.dart' as _i7;
+import 'package:firebase_auth/firebase_auth.dart' as _i13;
 import 'package:mockito/mockito.dart' as _i1;
 import 'package:projectunity/data/model/account/account.dart' as _i6;
 import 'package:projectunity/data/model/employee/employee.dart' as _i4;
 import 'package:projectunity/data/model/space/space.dart' as _i3;
 import 'package:projectunity/data/provider/device_info.dart' as _i5;
-import 'package:projectunity/data/provider/user_state.dart' as _i10;
-import 'package:projectunity/data/services/account_service.dart' as _i13;
-import 'package:projectunity/data/services/auth_service.dart' as _i14;
-import 'package:projectunity/data/services/employee_service.dart' as _i12;
-import 'package:projectunity/data/services/space_service.dart' as _i8;
+import 'package:projectunity/data/provider/user_state.dart' as _i9;
+import 'package:projectunity/data/services/account_service.dart' as _i12;
+import 'package:projectunity/data/services/employee_service.dart' as _i11;
+import 'package:projectunity/data/services/space_service.dart' as _i7;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
@@ -82,20 +81,10 @@ class _FakeAccount_4 extends _i1.SmartFake implements _i6.Account {
         );
 }
 
-class _FakeFirebaseAuth_5 extends _i1.SmartFake implements _i7.FirebaseAuth {
-  _FakeFirebaseAuth_5(
-    Object parent,
-    Invocation parentInvocation,
-  ) : super(
-          parent,
-          parentInvocation,
-        );
-}
-
 /// A class which mocks [SpaceService].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockSpaceService extends _i1.Mock implements _i8.SpaceService {
+class MockSpaceService extends _i1.Mock implements _i7.SpaceService {
   MockSpaceService() {
     _i1.throwOnMissingStub(this);
   }
@@ -117,15 +106,15 @@ class MockSpaceService extends _i1.Mock implements _i8.SpaceService {
         returnValueForMissingStub: null,
       );
   @override
-  _i9.Future<_i3.Space?> getSpace(String? spaceId) => (super.noSuchMethod(
+  _i8.Future<_i3.Space?> getSpace(String? spaceId) => (super.noSuchMethod(
         Invocation.method(
           #getSpace,
           [spaceId],
         ),
-        returnValue: _i9.Future<_i3.Space?>.value(),
-      ) as _i9.Future<_i3.Space?>);
+        returnValue: _i8.Future<_i3.Space?>.value(),
+      ) as _i8.Future<_i3.Space?>);
   @override
-  _i9.Future<_i3.Space> createSpace({
+  _i8.Future<_i3.Space> createSpace({
     String? logo,
     required String? name,
     String? domain,
@@ -144,7 +133,7 @@ class MockSpaceService extends _i1.Mock implements _i8.SpaceService {
             #ownerId: ownerId,
           },
         ),
-        returnValue: _i9.Future<_i3.Space>.value(_FakeSpace_1(
+        returnValue: _i8.Future<_i3.Space>.value(_FakeSpace_1(
           this,
           Invocation.method(
             #createSpace,
@@ -158,18 +147,18 @@ class MockSpaceService extends _i1.Mock implements _i8.SpaceService {
             },
           ),
         )),
-      ) as _i9.Future<_i3.Space>);
+      ) as _i8.Future<_i3.Space>);
   @override
-  _i9.Future<void> updateSpace(_i3.Space? space) => (super.noSuchMethod(
+  _i8.Future<void> updateSpace(_i3.Space? space) => (super.noSuchMethod(
         Invocation.method(
           #updateSpace,
           [space],
         ),
-        returnValue: _i9.Future<void>.value(),
-        returnValueForMissingStub: _i9.Future<void>.value(),
-      ) as _i9.Future<void>);
+        returnValue: _i8.Future<void>.value(),
+        returnValueForMissingStub: _i8.Future<void>.value(),
+      ) as _i8.Future<void>);
   @override
-  _i9.Future<void> deleteSpace(
+  _i8.Future<void> deleteSpace(
     String? workspaceId,
     List<String>? owners,
   ) =>
@@ -181,30 +170,30 @@ class MockSpaceService extends _i1.Mock implements _i8.SpaceService {
             owners,
           ],
         ),
-        returnValue: _i9.Future<void>.value(),
-        returnValueForMissingStub: _i9.Future<void>.value(),
-      ) as _i9.Future<void>);
+        returnValue: _i8.Future<void>.value(),
+        returnValueForMissingStub: _i8.Future<void>.value(),
+      ) as _i8.Future<void>);
   @override
-  _i9.Future<List<_i3.Space>> getSpacesOfUser(String? uid) =>
+  _i8.Future<List<_i3.Space>> getSpacesOfUser(String? uid) =>
       (super.noSuchMethod(
         Invocation.method(
           #getSpacesOfUser,
           [uid],
         ),
-        returnValue: _i9.Future<List<_i3.Space>>.value(<_i3.Space>[]),
-      ) as _i9.Future<List<_i3.Space>>);
+        returnValue: _i8.Future<List<_i3.Space>>.value(<_i3.Space>[]),
+      ) as _i8.Future<List<_i3.Space>>);
   @override
-  _i9.Future<int> getPaidLeaves({required String? spaceId}) =>
+  _i8.Future<int> getPaidLeaves({required String? spaceId}) =>
       (super.noSuchMethod(
         Invocation.method(
           #getPaidLeaves,
           [],
           {#spaceId: spaceId},
         ),
-        returnValue: _i9.Future<int>.value(0),
-      ) as _i9.Future<int>);
+        returnValue: _i8.Future<int>.value(0),
+      ) as _i8.Future<int>);
   @override
-  _i9.Future<void> updateLeaveCount({
+  _i8.Future<void> updateLeaveCount({
     required String? spaceId,
     required int? paidLeaveCount,
   }) =>
@@ -217,24 +206,24 @@ class MockSpaceService extends _i1.Mock implements _i8.SpaceService {
             #paidLeaveCount: paidLeaveCount,
           },
         ),
-        returnValue: _i9.Future<void>.value(),
-        returnValueForMissingStub: _i9.Future<void>.value(),
-      ) as _i9.Future<void>);
+        returnValue: _i8.Future<void>.value(),
+        returnValueForMissingStub: _i8.Future<void>.value(),
+      ) as _i8.Future<void>);
 }
 
 /// A class which mocks [UserStateNotifier].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockUserStateNotifier extends _i1.Mock implements _i10.UserStateNotifier {
+class MockUserStateNotifier extends _i1.Mock implements _i9.UserStateNotifier {
   MockUserStateNotifier() {
     _i1.throwOnMissingStub(this);
   }
 
   @override
-  _i10.UserState get state => (super.noSuchMethod(
+  _i9.UserState get state => (super.noSuchMethod(
         Invocation.getter(#state),
-        returnValue: _i10.UserState.unknown,
-      ) as _i10.UserState);
+        returnValue: _i9.UserState.unknown,
+      ) as _i9.UserState);
   @override
   String get employeeId => (super.noSuchMethod(
         Invocation.getter(#employeeId),
@@ -277,16 +266,16 @@ class MockUserStateNotifier extends _i1.Mock implements _i10.UserStateNotifier {
         returnValueForMissingStub: null,
       );
   @override
-  _i9.Future<void> setUser(_i6.Account? user) => (super.noSuchMethod(
+  _i8.Future<void> setUser(_i6.Account? user) => (super.noSuchMethod(
         Invocation.method(
           #setUser,
           [user],
         ),
-        returnValue: _i9.Future<void>.value(),
-        returnValueForMissingStub: _i9.Future<void>.value(),
-      ) as _i9.Future<void>);
+        returnValue: _i8.Future<void>.value(),
+        returnValueForMissingStub: _i8.Future<void>.value(),
+      ) as _i8.Future<void>);
   @override
-  _i9.Future<void> setEmployeeWithSpace({
+  _i8.Future<void> setEmployeeWithSpace({
     required _i3.Space? space,
     required _i4.Employee? spaceUser,
     bool? redirect = true,
@@ -301,38 +290,38 @@ class MockUserStateNotifier extends _i1.Mock implements _i10.UserStateNotifier {
             #redirect: redirect,
           },
         ),
-        returnValue: _i9.Future<void>.value(),
-        returnValueForMissingStub: _i9.Future<void>.value(),
-      ) as _i9.Future<void>);
+        returnValue: _i8.Future<void>.value(),
+        returnValueForMissingStub: _i8.Future<void>.value(),
+      ) as _i8.Future<void>);
   @override
-  _i9.Future<void> updateSpace(_i3.Space? space) => (super.noSuchMethod(
+  _i8.Future<void> updateSpace(_i3.Space? space) => (super.noSuchMethod(
         Invocation.method(
           #updateSpace,
           [space],
         ),
-        returnValue: _i9.Future<void>.value(),
-        returnValueForMissingStub: _i9.Future<void>.value(),
-      ) as _i9.Future<void>);
+        returnValue: _i8.Future<void>.value(),
+        returnValueForMissingStub: _i8.Future<void>.value(),
+      ) as _i8.Future<void>);
   @override
-  _i9.Future<void> removeEmployeeWithSpace() => (super.noSuchMethod(
+  _i8.Future<void> removeEmployeeWithSpace() => (super.noSuchMethod(
         Invocation.method(
           #removeEmployeeWithSpace,
           [],
         ),
-        returnValue: _i9.Future<void>.value(),
-        returnValueForMissingStub: _i9.Future<void>.value(),
-      ) as _i9.Future<void>);
+        returnValue: _i8.Future<void>.value(),
+        returnValueForMissingStub: _i8.Future<void>.value(),
+      ) as _i8.Future<void>);
   @override
-  _i9.Future<void> removeAll() => (super.noSuchMethod(
+  _i8.Future<void> removeAll() => (super.noSuchMethod(
         Invocation.method(
           #removeAll,
           [],
         ),
-        returnValue: _i9.Future<void>.value(),
-        returnValueForMissingStub: _i9.Future<void>.value(),
-      ) as _i9.Future<void>);
+        returnValue: _i8.Future<void>.value(),
+        returnValueForMissingStub: _i8.Future<void>.value(),
+      ) as _i8.Future<void>);
   @override
-  void addListener(_i11.VoidCallback? listener) => super.noSuchMethod(
+  void addListener(_i10.VoidCallback? listener) => super.noSuchMethod(
         Invocation.method(
           #addListener,
           [listener],
@@ -340,7 +329,7 @@ class MockUserStateNotifier extends _i1.Mock implements _i10.UserStateNotifier {
         returnValueForMissingStub: null,
       );
   @override
-  void removeListener(_i11.VoidCallback? listener) => super.noSuchMethod(
+  void removeListener(_i10.VoidCallback? listener) => super.noSuchMethod(
         Invocation.method(
           #removeListener,
           [listener],
@@ -368,7 +357,7 @@ class MockUserStateNotifier extends _i1.Mock implements _i10.UserStateNotifier {
 /// A class which mocks [EmployeeService].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockEmployeeService extends _i1.Mock implements _i12.EmployeeService {
+class MockEmployeeService extends _i1.Mock implements _i11.EmployeeService {
   MockEmployeeService() {
     _i1.throwOnMissingStub(this);
   }
@@ -382,7 +371,7 @@ class MockEmployeeService extends _i1.Mock implements _i12.EmployeeService {
         ),
       ) as _i2.FirebaseFirestore);
   @override
-  _i9.Future<void> addEmployeeBySpaceId({
+  _i8.Future<void> addEmployeeBySpaceId({
     required _i4.Employee? employee,
     required String? spaceId,
   }) =>
@@ -395,11 +384,11 @@ class MockEmployeeService extends _i1.Mock implements _i12.EmployeeService {
             #spaceId: spaceId,
           },
         ),
-        returnValue: _i9.Future<void>.value(),
-        returnValueForMissingStub: _i9.Future<void>.value(),
-      ) as _i9.Future<void>);
+        returnValue: _i8.Future<void>.value(),
+        returnValueForMissingStub: _i8.Future<void>.value(),
+      ) as _i8.Future<void>);
   @override
-  _i9.Future<_i4.Employee?> getEmployeeBySpaceId({
+  _i8.Future<_i4.Employee?> getEmployeeBySpaceId({
     required String? userId,
     required String? spaceId,
   }) =>
@@ -412,54 +401,54 @@ class MockEmployeeService extends _i1.Mock implements _i12.EmployeeService {
             #spaceId: spaceId,
           },
         ),
-        returnValue: _i9.Future<_i4.Employee?>.value(),
-      ) as _i9.Future<_i4.Employee?>);
+        returnValue: _i8.Future<_i4.Employee?>.value(),
+      ) as _i8.Future<_i4.Employee?>);
   @override
-  _i9.Future<List<_i4.Employee>> getEmployees() => (super.noSuchMethod(
+  _i8.Future<List<_i4.Employee>> getEmployees() => (super.noSuchMethod(
         Invocation.method(
           #getEmployees,
           [],
         ),
-        returnValue: _i9.Future<List<_i4.Employee>>.value(<_i4.Employee>[]),
-      ) as _i9.Future<List<_i4.Employee>>);
+        returnValue: _i8.Future<List<_i4.Employee>>.value(<_i4.Employee>[]),
+      ) as _i8.Future<List<_i4.Employee>>);
   @override
-  _i9.Future<_i4.Employee?> getEmployee(String? id) => (super.noSuchMethod(
+  _i8.Future<_i4.Employee?> getEmployee(String? id) => (super.noSuchMethod(
         Invocation.method(
           #getEmployee,
           [id],
         ),
-        returnValue: _i9.Future<_i4.Employee?>.value(),
-      ) as _i9.Future<_i4.Employee?>);
+        returnValue: _i8.Future<_i4.Employee?>.value(),
+      ) as _i8.Future<_i4.Employee?>);
   @override
-  _i9.Future<bool> hasUser(String? email) => (super.noSuchMethod(
+  _i8.Future<bool> hasUser(String? email) => (super.noSuchMethod(
         Invocation.method(
           #hasUser,
           [email],
         ),
-        returnValue: _i9.Future<bool>.value(false),
-      ) as _i9.Future<bool>);
+        returnValue: _i8.Future<bool>.value(false),
+      ) as _i8.Future<bool>);
   @override
-  _i9.Future<void> addEmployee(_i4.Employee? employee) => (super.noSuchMethod(
+  _i8.Future<void> addEmployee(_i4.Employee? employee) => (super.noSuchMethod(
         Invocation.method(
           #addEmployee,
           [employee],
         ),
-        returnValue: _i9.Future<void>.value(),
-        returnValueForMissingStub: _i9.Future<void>.value(),
-      ) as _i9.Future<void>);
+        returnValue: _i8.Future<void>.value(),
+        returnValueForMissingStub: _i8.Future<void>.value(),
+      ) as _i8.Future<void>);
   @override
-  _i9.Future<void> updateEmployeeDetails({required _i4.Employee? employee}) =>
+  _i8.Future<void> updateEmployeeDetails({required _i4.Employee? employee}) =>
       (super.noSuchMethod(
         Invocation.method(
           #updateEmployeeDetails,
           [],
           {#employee: employee},
         ),
-        returnValue: _i9.Future<void>.value(),
-        returnValueForMissingStub: _i9.Future<void>.value(),
-      ) as _i9.Future<void>);
+        returnValue: _i8.Future<void>.value(),
+        returnValueForMissingStub: _i8.Future<void>.value(),
+      ) as _i8.Future<void>);
   @override
-  _i9.Future<void> changeEmployeeRoleType(
+  _i8.Future<void> changeEmployeeRoleType(
     String? id,
     _i4.Role? role,
   ) =>
@@ -471,24 +460,24 @@ class MockEmployeeService extends _i1.Mock implements _i12.EmployeeService {
             role,
           ],
         ),
-        returnValue: _i9.Future<void>.value(),
-        returnValueForMissingStub: _i9.Future<void>.value(),
-      ) as _i9.Future<void>);
+        returnValue: _i8.Future<void>.value(),
+        returnValueForMissingStub: _i8.Future<void>.value(),
+      ) as _i8.Future<void>);
   @override
-  _i9.Future<void> deleteEmployee(String? id) => (super.noSuchMethod(
+  _i8.Future<void> deleteEmployee(String? id) => (super.noSuchMethod(
         Invocation.method(
           #deleteEmployee,
           [id],
         ),
-        returnValue: _i9.Future<void>.value(),
-        returnValueForMissingStub: _i9.Future<void>.value(),
-      ) as _i9.Future<void>);
+        returnValue: _i8.Future<void>.value(),
+        returnValueForMissingStub: _i8.Future<void>.value(),
+      ) as _i8.Future<void>);
 }
 
 /// A class which mocks [AccountService].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockAccountService extends _i1.Mock implements _i13.AccountService {
+class MockAccountService extends _i1.Mock implements _i12.AccountService {
   MockAccountService() {
     _i1.throwOnMissingStub(this);
   }
@@ -510,21 +499,21 @@ class MockAccountService extends _i1.Mock implements _i13.AccountService {
         ),
       ) as _i5.DeviceInfoProvider);
   @override
-  _i9.Future<_i6.Account> getUser(_i7.User? authData) => (super.noSuchMethod(
+  _i8.Future<_i6.Account> getUser(_i13.User? authData) => (super.noSuchMethod(
         Invocation.method(
           #getUser,
           [authData],
         ),
-        returnValue: _i9.Future<_i6.Account>.value(_FakeAccount_4(
+        returnValue: _i8.Future<_i6.Account>.value(_FakeAccount_4(
           this,
           Invocation.method(
             #getUser,
             [authData],
           ),
         )),
-      ) as _i9.Future<_i6.Account>);
+      ) as _i8.Future<_i6.Account>);
   @override
-  _i9.Future<void> updateSpaceOfUser({
+  _i8.Future<void> updateSpaceOfUser({
     required String? spaceID,
     required String? uid,
   }) =>
@@ -537,11 +526,11 @@ class MockAccountService extends _i1.Mock implements _i13.AccountService {
             #uid: uid,
           },
         ),
-        returnValue: _i9.Future<void>.value(),
-        returnValueForMissingStub: _i9.Future<void>.value(),
-      ) as _i9.Future<void>);
+        returnValue: _i8.Future<void>.value(),
+        returnValueForMissingStub: _i8.Future<void>.value(),
+      ) as _i8.Future<void>);
   @override
-  _i9.Future<void> deleteSpaceIdFromAccount({
+  _i8.Future<void> deleteSpaceIdFromAccount({
     required String? spaceId,
     required String? uid,
   }) =>
@@ -554,59 +543,17 @@ class MockAccountService extends _i1.Mock implements _i13.AccountService {
             #uid: uid,
           },
         ),
-        returnValue: _i9.Future<void>.value(),
-        returnValueForMissingStub: _i9.Future<void>.value(),
-      ) as _i9.Future<void>);
+        returnValue: _i8.Future<void>.value(),
+        returnValueForMissingStub: _i8.Future<void>.value(),
+      ) as _i8.Future<void>);
   @override
-  _i9.Future<List<String>> fetchSpaceIds({required String? uid}) =>
+  _i8.Future<List<String>> fetchSpaceIds({required String? uid}) =>
       (super.noSuchMethod(
         Invocation.method(
           #fetchSpaceIds,
           [],
           {#uid: uid},
         ),
-        returnValue: _i9.Future<List<String>>.value(<String>[]),
-      ) as _i9.Future<List<String>>);
-}
-
-/// A class which mocks [AuthService].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockAuthService extends _i1.Mock implements _i14.AuthService {
-  MockAuthService() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  _i2.FirebaseFirestore get fireStore => (super.noSuchMethod(
-        Invocation.getter(#fireStore),
-        returnValue: _FakeFirebaseFirestore_0(
-          this,
-          Invocation.getter(#fireStore),
-        ),
-      ) as _i2.FirebaseFirestore);
-  @override
-  _i7.FirebaseAuth get firebaseAuth => (super.noSuchMethod(
-        Invocation.getter(#firebaseAuth),
-        returnValue: _FakeFirebaseAuth_5(
-          this,
-          Invocation.getter(#firebaseAuth),
-        ),
-      ) as _i7.FirebaseAuth);
-  @override
-  _i9.Future<_i7.User?> signInWithGoogle() => (super.noSuchMethod(
-        Invocation.method(
-          #signInWithGoogle,
-          [],
-        ),
-        returnValue: _i9.Future<_i7.User?>.value(),
-      ) as _i9.Future<_i7.User?>);
-  @override
-  _i9.Future<bool> signOutWithGoogle() => (super.noSuchMethod(
-        Invocation.method(
-          #signOutWithGoogle,
-          [],
-        ),
-        returnValue: _i9.Future<bool>.value(false),
-      ) as _i9.Future<bool>);
+        returnValue: _i8.Future<List<String>>.value(<String>[]),
+      ) as _i8.Future<List<String>>);
 }
